refactor: more pertinent naming and removing redundant code block
`extract` was too much of an ambiguous
function name so for the sake of clarity
I renamed it to `constructHuffmanTree`.
Also, I noticed a remnant of a bygone
era, which got me scratching my head
because I had no idea why I wrote it. It
is redundant so I removed it